pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rube-iks-cube"
version = "0.1.0"
description = "MCP client for rube.app"
dependencies = [
    "pydantic-ai>=0.7.4",
    "rich>=13.0.0",
    "composio-core[all]>=0.7.20",
    "python-dotenv>=1.1.1",
]
requires-python = ">=3.10"

[tool.setuptools]
py-modules = ["main"]

